+++ client/src/components/atoms/SearchBar.tsx
