+++ client/src/components/molecules/PlaceInfo.tsx
